pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orglens-agent"
version = "0.1.0"
description = "OrgLens Layer 1 — Local data collection agent for GitHub repositories"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "perceval>=0.21",
    "perceval-mozilla>=0.3",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "httpx>=0.27",
    "PyYAML>=6.0",
    "APScheduler>=3.10",
    "asyncpg>=0.29",
    "redis>=5.0",
    "aioboto3>=13.1",
    "pyarrow>=17.0",
    "python-dateutil>=2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "httpx",          # also used in tests for TestClient
    "ruff>=0.4",
]

[project.scripts]
orglens-agent = "orglens.layers.layer1.agent:main"
orglens-cloud-ingest = "orglens.layers.layer1_cloud.main:main"
orglens-core = "orglens.layers.layer2_core.main:main"
orglens-pipeline = "orglens.pipeline:main"
orglens-auto = "orglens.auto_cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["orglens*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = [
    "data/repos/**",
    "archive/**",
    "reports/archive/**",
]
